+++ ChangeLog.md
@@ -9,6 +9,7 @@ Added in 0.9.1
     * `Match.group().textLength()`
     * `Match.groupsCount()`
     * Add method `by()->group()->callback()` which accepts `MatchGroup` as an argument
+    * Method `by()->group()->orElse()` now receives lazy-loaded `Match`, instead of a subject
 
 Available in 0.9.0
 ---------------